refactor(settings): relocate connection status from settings to Actions
The Connectivity settings tab held no editable setting once routing and
subscription became fixed — only status: SSO session identity, server/stream
state, versions, internet reachability and a delta-throughput chart. Move that
diagnostics component out of the settings tab group into its own
connection-status component, reached by a new "Connection" tile on the Actions
page (mirroring Settings / Remote Control / Help).

Refs #385
