+++ src/app/core/components/connection-status/connection-status.component.html
@@ -1,0 +1,30 @@
+<div class="sso-identity" aria-live="polite">
+  @if (isUserSession()) {
+    <span>Signed in@if (loginStatus()?.oidcProviderName) { via {{ loginStatus()?.oidcProviderName }}}@if (loginStatus()?.username) { as {{ loginStatus()?.username }}}.</span>
+    @if (!canWriteUserData()) {
+      <span class="sso-readonly"> Read-only access — configuration changes are disabled on the server.</span>
+    }
+  } @else if (loginStatus(); as status) {
+    @if (status.authenticationRequired === false) {
+      <span>Connected (no sign-in required).</span>
+    } @else {
+      <span>Not signed in. </span>
+      <button mat-flat-button type="button" (click)="signIn()">Sign in</button>
+    }
+  } @else {
+    <span>Checking sign-in status…</span>
+  }
+</div>
+
+<div class="activity-container">
+  <div class="activity-graph-container">
+    <canvas class="activity-graph" #activityGraph></canvas>
+  </div>
+  <div class="info-container">
+    <pre>Skip: {{ app.appVersion() }}
+{{ app.browserVersion() }}, {{ app.osVersion() }}<br />
+{{ endpointServiceStatus().serverDescription }}
+{{ streamStatus().message }}, {{ isUserSession() ? 'Signed in' : 'Not signed in' }}
+Internet: {{ internetAvailabilityLabel() }}</pre>
+  </div>
+</div>
