--- src/app/core/components/settings/signalk/signalk.component.html
@@ -1,62 +1,0 @@
-<form name="connectionForm" (ngSubmit)="connectToServer()" #connectionForm="ngForm">
-  <br />
-  <table class="connect-table">
-    <tr class="connect-table">
-      <td class="url-cell">
-        <mat-form-field class="connectUrlInput">
-          <mat-label>Signal K URL</mat-label>
-          <input type=url matInput placeholder="Ex. http://192.168.0.1:3000" name="signalKURL" [(ngModel)]="connectionConfig.signalKUrl" [ngModelOptions]="{standalone: false}" required pattern="https?://.+">
-          <mat-error>
-            Valid URL is required. Ex. "https://demo.signalK.org" or "http://my.server.com:3000"
-          </mat-error>
-        </mat-form-field>
-      </td>
-      <td class="connect-cell">
-        <button mat-flat-button class="connect-btn" type="submit" [disabled]="!connectionForm.form.valid || isConnecting()">
-          {{ isConnecting() ? 'Validating...' : 'Connect' }}
-        </button>
-      </td>
-    </tr>
-    <tr class="connect-table">
-      <td colspan="2">
-        <div class="sso-identity" aria-live="polite">
-          @if (isUserSession()) {
-            <span>Signed in@if (loginStatus()?.oidcProviderName) { via {{ loginStatus()?.oidcProviderName }}}@if (loginStatus()?.username) { as {{ loginStatus()?.username }}}.</span>
-            @if (!canWriteUserData()) {
-              <span class="sso-readonly"> Read-only access — configuration changes are disabled on the server.</span>
-            }
-          } @else if (loginStatus(); as status) {
-            @if (status.authenticationRequired === false) {
-              <span>Connected (no sign-in required).</span>
-            } @else {
-              <span>Not signed in. </span>
-              <button mat-flat-button type="button" (click)="signIn()">Sign in</button>
-            }
-          } @else {
-            <span>Checking sign-in status…</span>
-          }
-        </div>
-        <br />
-        <mat-checkbox [(ngModel)]="connectionConfig.signalKSubscribeAll" [ngModelOptions]="{standalone: false}" name="signalKSubscribeAll">Subscribe to remote sources messages such as AIS and DSC targets. <i>Warning: this maybe generate heavy traffic.</i></mat-checkbox>
-      </td>
-    </tr>
-    <tr class="connect-table">
-      <td colspan="2">
-        <mat-checkbox [(ngModel)]="connectionConfig.proxyEnabled" [ngModelOptions]="{standalone: false}" name="isProxyEnabled">Signal K server is behind a Proxy</mat-checkbox>
-      </td>
-    </tr>
-  </table>
-</form>
-
-  <div class="activity-container">
-    <div class="activity-graph-container">
-      <canvas class="activity-graph" #activityGraph></canvas>
-    </div>
-    <div class="info-container">
-      <pre>Skip: {{ app.appVersion() }}
-{{ app.browserVersion() }}, {{ app.osVersion() }}<br />
-{{ endpointServiceStatus().serverDescription }}
-{{ streamStatus().message }}, {{ isUserSession() ? 'Signed in' : 'Not signed in' }}
-Internet: {{ internetAvailabilityLabel() }}</pre>
-    </div>
-  </div>
