+++ src/app/core/components/settings/tabs/tabs.component.html
@@ -1,8 +1,5 @@
 <page-header [pageTitle]="pageTitle" svgIconId="settings" />
 <mat-tab-group class="content-area">
-  <mat-tab label="Connectivity">
-      <settings-signalk class="tab-content" />
-  </mat-tab>
   <mat-tab label="Display">
     <settings-display class="tab-content" />
   </mat-tab>
